Split out partial liquidation tests (#2113)
* Split out partial liquidation tests

* Make sure we wait for configs to be mined

* Make sure we mine flagging

* See if explicitly mining liqs in same blocks helps
